% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_single_gene.R
\name{simulateCisEffect.s2m}
\alias{simulateCisEffect.s2m}
\title{Simulate cis-regulatory effect for one-to-multi gene-snp pairs}
\usage{
simulateCisEffect.s2m(
  n_i,
  n_k,
  snp_pars,
  phi,
  prob_as,
  theta,
  baseline,
  origin = NULL,
  origin.effect = NULL,
  lib.size = NULL,
  confounder = NULL,
  confounder.effect = NULL
)
}
\arguments{
\item{n_i}{Number of samples}

\item{n_k}{Number of regulatory snps}

\item{snp_pars}{A list of length \code{n_k} or \code{1} specifying
generative parameters for each regulatory snp. Each element should be a
list containing parameters used in \code{\link{simulateGenotype.s2s}},
\code{r}, the cis-regulatory effect (log read ratio), and \code{p_error},
a vector of phasing error rate of the gene pair in each sample.
If \code{length(p_error) == 1} then all samples share the same
phasing error rate.
If \code{length(snp_pars) == 1} then all \code{n_k} snps share the same set
of parameters.}

\item{phi}{Overdispersion term in negative-binomial}

\item{prob_as}{Proportion of allele-specific reads}

\item{theta}{Overdispersion term in beta-binomial}

\item{baseline}{Baseline expression (under ref/ref scenario)}

\item{origin}{A vector specifying sample origin. If \code{NULL} then no two
samples come from a same individual.}

\item{origin.effect}{A vector specifying the effect of sample origin with
length \code{n_indiv} (the number of samples). If \code{NULL} then no effect
will be added}

\item{lib.size}{A vector of the relative library size of each sample. If
\code{NULL} then no adjustment will be made}

\item{confounder}{A n_i by N matrix where N is the number of confounding factors}

\item{confounder.effect}{A vector specifying the effect of each confounding factor
If \code{NULL} then no effect will be added}
}
\description{
This function simulates genotype G, phasing P, and expression
(total and allele-specific) profiles for one-to-multi gene-snp pairs
}
\details{
J=1, K>=1, L=1
}
