% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_single_gene.R
\name{simulateCisEffect.s2s}
\alias{simulateCisEffect.s2s}
\title{Simulate cis-regulatory effect for one gene-snp pair}
\usage{
simulateCisEffect.s2s(
  n_i,
  maf,
  prob_ref,
  phi,
  prob_as,
  theta,
  baseline,
  r,
  p_error = NULL
)
}
\arguments{
\item{n_i}{Number of samples}

\item{maf}{Minor allele frequency of the test snp}

\item{prob_ref}{Probability of the test snp on the ref allele of the target
gene if the test snp is heterogeneous}

\item{phi}{Overdispersion term in negative-binomial}

\item{prob_as}{Proportion of allele-specific reads}

\item{theta}{Overdispersion term in beta-binomial}

\item{baseline}{Baseline expression (under ref/ref scenario)}

\item{r}{Cis-regulatory effect defined as the log read ratio under alt/alt
and ref/ref scenario}

\item{p_error}{A vector of phasing error rate of the gene pair in each
sample. If \code{length(p_error) == 1} then all samples share the same
phasing error rate}
}
\description{
This function simulates genotype G, phasing P, and expression
(total and allele-specific) profiles for one gene-snp pair
}
\details{
J=1, K=1, L=1
}
