% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateCisEffectMulti}
\alias{simulateCisEffectMulti}
\title{Simulate cis-regulatory effect for multi-to-one gene-snp pairs}
\usage{
simulateCisEffectMulti(n_i, n_j, maf, prob_ref, gene_pars)
}
\arguments{
\item{n_i}{Numbet of samples}

\item{n_j}{Number of genes}

\item{maf}{Minor allele frequency of the test snp}

\item{prob_ref}{A vector specifying the probability of the test snp on the
ref allele of each gene if the test snp is heterogeneous.
If \code{length(prob_ref) == 1} then all genes share the same \code{prob_ref}}

\item{gene_pars}{A vector of length \code{n_j} or \code{1} specifying
gene-level statistics. Each element should be a list containing parameters
used in \code{\link{simulateCisEffectSingle}}, including phi, prob_as, theta,
baseline, and r. If \code{length(gene_pars) == 1} then all \code{n_j} genes
share the same set of parameters.}
}
\description{
This function simulates genotype G, phasing P, and expression
(total and allele-specific) profiles for multi-to-one gene-snp pairs
}
\details{
J>=1, K=1, L=1
}
