% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_multi_genes.R
\name{simulateGenotype.m2s}
\alias{simulateGenotype.m2s}
\title{Simulate genotype for multi-to-one gene-snp pairs}
\usage{
simulateGenotype.m2s(n_i, n_j, maf, prob_ref)
}
\arguments{
\item{n_i}{Number of samples}

\item{n_j}{Number of genes}

\item{maf}{Minor allele frequency of the test snp}

\item{prob_ref}{A vector specifying the probability of the test snp on the
ref allele of each gene if the test snp is heterogeneous.
If \code{length(prob_ref) == 1} then all genes share the same \code{prob_ref}}
}
\description{
This function simulates genotype G and phasing P for multi-to-one gene-snp pairs
}
\details{
J>=1, K=1, L=1
}
