% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_single_gene.R
\name{simulateGenotype.s2m}
\alias{simulateGenotype.s2m}
\title{Simulate genotype for one-to-multi gene-snp pairs}
\usage{
simulateGenotype.s2m(n_i, n_k, snp_pars, origin = NULL)
}
\arguments{
\item{n_i}{Number of samples}

\item{n_k}{Number of regulatory snps}

\item{snp_pars}{A list of length \code{n_k} or \code{1} specifying
generative parameters for each regulatory snp. Each element should be a
list containing parameters used in \code{\link{simulateGenotype.s2s}}.
If \code{length(snp_pars) == 1} then all \code{n_k} snps share the same set
of parameters.}

\item{origin}{A vector specifying sample origin. If \code{NULL} then no two
samples come from a same individual.}
}
\description{
This function simulates genotype G and phasing P for one-to-multi gene-snp pairs
}
\details{
J=1, K>=1, L=1
}
