% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_single_gene.R
\name{simulateGenotype.s2s}
\alias{simulateGenotype.s2s}
\title{Simulate genotype for one gene-snp pair}
\usage{
simulateGenotype.s2s(n_i, maf, prob_ref, origin = NULL)
}
\arguments{
\item{n_i}{Number of samples}

\item{maf}{Minor allele frequency of the test snp}

\item{prob_ref}{Probability of the test snp on the ref allele of the target
gene if the test snp is heterogeneous}

\item{origin}{A vector specifying sample origin. If \code{NULL} then no two
samples come from a same individual.}
}
\description{
This function simulates genotype G and phasing P for one gene-snp pair
}
\details{
J=1, K=1, L=1
}
