% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateGenotypeSingle}
\alias{simulateGenotypeSingle}
\title{Simulate genotype for one gene-snp pair}
\usage{
simulateGenotypeSingle(n_i, maf, prob_ref)
}
\arguments{
\item{n_i}{Number of samples}

\item{maf}{Minor allele frequency of the test snp}

\item{prob_ref}{Probability of the test snp on the ref allele of the target
gene if the test snp is heterogeneous}
}
\description{
This function simulates genotype G and phasing P for one gene-snp pair
}
\details{
J=1, K=1, L=1
}
