% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_single_gene.R
\name{simulateReadCounts.s}
\alias{simulateReadCounts.s}
\title{Simulate read for one gene}
\usage{
simulateReadCounts.s(n_i, mu, phi, prob_as, prob_alt, theta)
}
\arguments{
\item{n_i}{Number of samples}

\item{mu}{Expected number of total reads}

\item{phi}{Overdispersion term in negative-binomial}

\item{prob_as}{Proportion of allele-specific reads}

\item{prob_alt}{Proportion of alt-allele reads in allele-specific reads}

\item{theta}{Overdispersion term in beta-binomial}
}
\description{
This function simulates read counts data from negative binomial (total read counts)
and beta-binomial (allele-specific read counts) distributions for one
}
\details{
J=1, K=1, L=1

\code{T ~ NB(mu, phi)}
\code{A_alt ~ BB(T * prob_as := N_as, prob_alt, theta)}
}
