% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateReadCountsMulti}
\alias{simulateReadCountsMulti}
\title{Simulate read for multiple genes}
\usage{
simulateReadCountsMulti(n_i, n_j, gene_pars)
}
\arguments{
\item{n_i}{Numbet of samples}

\item{n_j}{Number of genes}

\item{gene_pars}{A list of length \code{n_j} or \code{1} specifying
gene-level statistics. Each element should be a list containing parameters
used in \code{\link{simulateReadCountsSingle}}.
If \code{length(gene_pars) == 1} then all \code{n_j} genes share the same set
of parameters.}
}
\description{
This function simulates read counts data from negative binomial (total read counts)
and beta-binomial (allele-specific read counts) distributions for multiple genes
}
\details{
J>=1, K=1, L=1

\code{T ~ NB(mu, phi)}
\code{A_alt ~ BB(T * prob_as := N_as, prob_alt, theta)}
}
